+++ src/main/java/com/seodisparate/TurnBasedMinecraft/common/EditingInfo.java
@@ -7,12 +7,14 @@ public class EditingInfo
     public EntityPlayer editor;
     public EntityInfo entityInfo;
     public boolean isPendingEntitySelection;
+    public boolean isEditingCustomName;
 
     public EditingInfo()
     {
         editor = null;
         entityInfo = null;
         isPendingEntitySelection = true;
+        isEditingCustomName = false;
     }
 
     public EditingInfo(EntityPlayer player)
@@ -20,5 +22,6 @@ public class EditingInfo
         editor = player;
         entityInfo = null;
         isPendingEntitySelection = true;
+        isEditingCustomName = false;
     }
 }